# FFM-11115 - Update Evaluation logic to be more efficient