@stephenmcconkey 2024-04-04
Tests passed on Testgrid running locally.

Holding off on merge until next week

<img width="2159" alt="Screenshot 2024-04-04 at 16 34 56" src="https://github.com/harness/ff-golang-server-sdk/assets/62011008/af2625c1-9337-4d65-8422-deaf33baac1b">